pyproject.toml
[project]
name = "bot-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.3",
    "pynacl>=1.6.0",
    "python-dotenv>=1.1.1",
    "yt-dlp>=2025.9.5",
]
